parfor alloc checks

+++ benchmarks/CliqueCounting/intersect.h
@@ -57,7 +57,7 @@ struct HybridSpace_lw {
   HybridSpace_lw () {}
 
   void alloc(size_t max_induced, size_t k, size_t n, bool _use_old_labels, bool _use_base, bool _free_relabel=true) {
-    bool checking = pbbslib::CAS(&worker_in_use, UINT_E_MAX, worker_id());
+    bool checking = pbbslib::CAS(&worker_in_use, static_cast<size_t>(UINT_E_MAX), static_cast<size_t>(worker_id()));
     if (!checking) {
       std::cout << "ParForAlloc Thread err" << std::endl; fflush(stdout);
     }
